Move descriptor module and DescriptorSetDesc into layout module
