+++ vulkano/src/command_buffer/sys.rs
@@ -17,7 +17,7 @@ use crate::command_buffer::CommandBufferLevel;
 use crate::command_buffer::CommandBufferUsage;
 use crate::command_buffer::SecondaryCommandBuffer;
 use crate::command_buffer::SubpassContents;
-use crate::descriptor_set::descriptor::ShaderStages;
+use crate::descriptor_set::layout::ShaderStages;
 use crate::descriptor_set::sys::UnsafeDescriptorSet;
 use crate::device::Device;
 use crate::device::DeviceOwned;
