+++ vulkano/src/descriptor_set/layout/sys.rs
@@ -8,7 +8,7 @@
 // according to those terms.
 
 use crate::check_errors;
-use crate::descriptor_set::descriptor::DescriptorDesc;
+use crate::descriptor_set::layout::DescriptorDesc;
 use crate::descriptor_set::pool::DescriptorsCount;
 use crate::descriptor_set::DescriptorSetDesc;
 use crate::device::Device;
@@ -159,12 +159,12 @@ impl Drop for DescriptorSetLayout {
 
 #[cfg(test)]
 mod tests {
-    use crate::descriptor_set::descriptor::DescriptorBufferDesc;
-    use crate::descriptor_set::descriptor::DescriptorDesc;
-    use crate::descriptor_set::descriptor::DescriptorDescTy;
-    use crate::descriptor_set::descriptor::ShaderStages;
+    use crate::descriptor_set::layout::DescriptorBufferDesc;
+    use crate::descriptor_set::layout::DescriptorDesc;
+    use crate::descriptor_set::layout::DescriptorDescTy;
+    use crate::descriptor_set::layout::DescriptorSetLayout;
+    use crate::descriptor_set::layout::ShaderStages;
     use crate::descriptor_set::pool::DescriptorsCount;
-    use crate::descriptor_set::DescriptorSetLayout;
     use std::iter;
 
     #[test]
