+++ vulkano/src/descriptor_set/pool/mod.rs
@@ -11,8 +11,8 @@ pub use self::standard::StdDescriptorPool;
 pub use self::sys::DescriptorPoolAllocError;
 pub use self::sys::UnsafeDescriptorPool;
 pub use self::sys::UnsafeDescriptorPoolAllocIter;
-use crate::descriptor_set::descriptor::DescriptorType;
-use crate::descriptor_set::DescriptorSetLayout;
+use crate::descriptor_set::layout::DescriptorSetLayout;
+use crate::descriptor_set::layout::DescriptorType;
 use crate::descriptor_set::UnsafeDescriptorSet;
 use crate::device::DeviceOwned;
 use crate::OomError;
